requirements.txt
customtkinter
Pillow
packaging